requirements.txt
# vim:fileencoding=utf-8:foldmethod=marker

# System {{{
fastapi[standard]
uvicorn
slowapi
python-dotenv
Jinja2
SQLAlchemy
python-multipart
aiofiles
apscheduler
# vercel_blob
pymysql
PyMySQL[rsa]
# }}}

# userCode {{{
redis
slowapi
httpx
# apscheduler

# vercel_blob
# types-redis
# tkinter
# shutil
# pprint

# }}}


